StringConcatenation
String Concatenation problem solved using regex. Other method involves a loop and it may pass more test cases but we'll see. The problem is as follows:

The authority of XYZ gated residential colony wants its residents' name datum Should be stored in the following format - residents' name <space> his/her father's name. Write a program to concat the father's name to the residents' name. The name should be validated,on validation the name should contain only alphabets and space is allowed. If the name is not valid display the message "Invalid name". If valid string then convert  it to uppercase and print it.