Student
Goes along with StudentMain file. Invokes two student constructors with the following parameters:
If student belongs to NIT, give input as 'yes/YES' and  skip input for the attribute collegeName  and create student object with 3-argument constructor to initilze the values for studentId, studentName and studentAddress and  collegeName as “NIT”.
If student belongs to other college, give input as 'no/NO' and get college name from the user and create student object with 4-argument constructor to initialize all the values. 
Instead of Yes / No, if user enters different input then display 'Wrong Input' and get the input again.